Finalising the split of the utilities file
